# api/openapi.yaml
openapi: 3.1.0
info:
  title: bili-notify administration API
  version: 3.0.0
  description: |
    Authoritative transport contract for the same-origin TLS administration UI.
    Version 3 unifies Bilibili and Knowledge Planet accounts, sources, content and comment trees.
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
  /api/v3/session:
    get:
      tags: [session]
      security: []
      operationId: getSession
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [session]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/CSRFResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [session]
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Session deleted.
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/CSRFError'
  /api/v3/setup:
    post:
      tags: [session]
      security: []
      operationId: setupAdministrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '200':
          $ref: '#/components/responses/CSRFResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v3/session/password:
    put:
      tags: [session]
      operationId: changePassword
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/CSRFResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/CSRFError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v3/settings:
    get:
      tags: [runtime]
      operationId: getSettings
      responses:
        '200':
          $ref: '#/components/responses/SettingsResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
    put:
      tags: [runtime]
      operationId: updateSettings
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeSettings'
      responses:
        '200':
          $ref: '#/components/responses/SettingsResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/CSRFError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v3/ws:
    get:
      tags: [runtime]
      operationId: connectRealtime
      description: |
        WebSocket upgrade endpoint. The server first sends `sync.required` and
        subsequently sends `resources.invalidated`. Messages never contain resource data.
      responses:
        '101':
          description: WebSocket upgraded. Messages conform to RealtimeEvent.
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/CSRFError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bili_notify_session
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 1}
...
  responses:
    SessionResponse:
      description: Current initialization and authentication state.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Session'
    CSRFResponse:
      description: Authenticated session created or replaced.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Set-Cookie:
          schema: {type: string}
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CSRFToken'
    SettingsResponse:
      description: Complete runtime settings.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RuntimeSettings'
...
    InvalidRequest:
      description: Invalid JSON, query, cursor, or validation failure.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    AuthenticationError:
      description: Authentication required or credentials invalid.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    CSRFError:
      description: CSRF token or WebSocket origin rejected.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ConflictError:
      description: Resource state conflicts with the operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimitError:
      description: Authentication attempts are rate limited.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Internal server failure.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
    Session:
      type: object
      additionalProperties: false
      required: [setup_required, authenticated]
      properties:
        setup_required: {type: boolean}
        authenticated: {type: boolean}
        csrf_token: {type: string}
    CSRFToken:
      type: object
      additionalProperties: false
      required: [csrf_token]
      properties:
        csrf_token: {type: string, minLength: 1}
    SetupRequest:
      type: object
      additionalProperties: false
      required: [setup_code, password]
      properties:
        setup_code: {type: string, minLength: 1}
        password: {type: string, minLength: 12}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password: {type: string}
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password: {type: string}
        new_password: {type: string, minLength: 12}
...
    RuntimeSettings:
      type: object
      additionalProperties: false
      required:
        - bilibili_dynamic_interval_sec
        - bilibili_request_rate
        - bilibili_request_concurrency
        - bilibili_comments_enabled
        - bilibili_comment_track_n
        - bilibili_comment_interval_sec
        - bilibili_relation_refresh_interval_sec
        - bilibili_space_reconcile_interval_sec
        - bilibili_max_dynamic_pages
        - bilibili_risk_pause_sec
        - zsxq_dynamic_interval_sec
        - zsxq_comment_interval_sec
        - zsxq_comments_enabled
        - zsxq_request_rate
        - zsxq_request_concurrency
        - zsxq_risk_pause_sec
        - zsxq_asset_max_file_mib
        - zsxq_asset_total_budget_gib
        - log_level
        - audit_log_retention_days
        - delivery_concurrency
        - backlog_alert_count
        - backlog_alert_age_sec
        - delivery_retry_delays_sec
        - ai_auto_processing_enabled
      properties:
        bilibili_dynamic_interval_sec: {type: integer, minimum: 10, maximum: 86400}
        bilibili_request_rate: {type: number, exclusiveMinimum: 0, maximum: 10}
        bilibili_request_concurrency: {type: integer, minimum: 1, maximum: 16}
        bilibili_comments_enabled: {type: boolean}
        bilibili_comment_track_n: {type: integer, minimum: 1, maximum: 50}
        bilibili_comment_interval_sec: {type: integer, minimum: 30, maximum: 86400}
        bilibili_relation_refresh_interval_sec: {type: integer, minimum: 60, maximum: 86400}
        bilibili_space_reconcile_interval_sec: {type: integer, minimum: 300, maximum: 604800}
        bilibili_max_dynamic_pages: {type: integer, minimum: 1, maximum: 20}
        bilibili_risk_pause_sec: {type: integer, minimum: 60, maximum: 3600}
        zsxq_dynamic_interval_sec: {type: integer, minimum: 30, maximum: 86400}
        zsxq_comment_interval_sec: {type: integer, minimum: 30, maximum: 86400}
        zsxq_comments_enabled: {type: boolean}
        zsxq_request_rate: {type: number, exclusiveMinimum: 0, maximum: 10}
        zsxq_request_concurrency: {type: integer, minimum: 1, maximum: 16}
        zsxq_risk_pause_sec: {type: integer, minimum: 60, maximum: 3600}
        zsxq_asset_max_file_mib: {type: integer, minimum: 1, maximum: 2048}
        zsxq_asset_total_budget_gib: {type: integer, minimum: 1, maximum: 10240}
        log_level: {type: string, enum: [debug, info, warn, error]}
        audit_log_retention_days: {type: integer, minimum: 1, maximum: 3650}
        delivery_concurrency: {type: integer, minimum: 1, maximum: 32}
        backlog_alert_count: {type: integer, minimum: 1, maximum: 100000}
        backlog_alert_age_sec: {type: integer, minimum: 60, maximum: 86400}
        delivery_retry_delays_sec:
          type: array
          description: Five retry delays in nondecreasing order.
          minItems: 5
          maxItems: 5
          items: {type: integer, minimum: 1, maximum: 86400}
        ai_auto_processing_enabled:
          type: boolean
          description: Automatically transcribe and summarize newly discovered top-level video dynamics using the enabled default AI profiles and prompt.
...